Update spring and move some bug tests to spring package for easier exclusion
