+++ integrationtest/src/test/java/org/mapstruct/itest/tests/FullFeatureCompilationExclusionCliEnhancer.java
@@ -30,10 +30,7 @@ public final class FullFeatureCompilationExclusionCliEnhancer implements Process
 
         switch ( currentJreVersion ) {
             case JAVA_8:
-                additionalExcludes.add( "org/mapstruct/ap/test/*/spring/**/*.java" );
-                additionalExcludes.add( "org/mapstruct/ap/test/bugs/_880/**/*.java" );
-                additionalExcludes.add( "org/mapstruct/ap/test/bugs/_1395/**/*.java" );
-                additionalExcludes.add( "org/mapstruct/ap/test/bugs/_2807/**/*.java" );
+                additionalExcludes.add( "org/mapstruct/ap/test/**/spring/**/*.java" );
                 additionalExcludes.add( "org/mapstruct/ap/test/injectionstrategy/cdi/**/*.java" );
                 additionalExcludes.add( "org/mapstruct/ap/test/injectionstrategy/jakarta_cdi/**/*.java" );
                 additionalExcludes.add( "org/mapstruct/ap/test/annotatewith/deprecated/jdk11/*.java" );
@@ -47,10 +44,7 @@ public final class FullFeatureCompilationExclusionCliEnhancer implements Process
                 additionalExcludes.add( "org/mapstruct/ap/test/collection/wildcard/BeanMapper.java" );
                 break;
             case JAVA_11:
-                additionalExcludes.add( "org/mapstruct/ap/test/*/spring/**/*.java" );
-                additionalExcludes.add( "org/mapstruct/ap/test/bugs/_880/**/*.java" );
-                additionalExcludes.add( "org/mapstruct/ap/test/bugs/_1395/**/*.java" );
-                additionalExcludes.add( "org/mapstruct/ap/test/bugs/_2807/**/*.java" );
+                additionalExcludes.add( "org/mapstruct/ap/test/**/spring/**/*.java" );
             default:
         }
