+++ processor/src/test/java/org/mapstruct/ap/test/bugs/_1395/spring/Source.java
@@ -3,12 +3,12 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.bugs._1395;
+package org.mapstruct.ap.test.bugs._1395.spring;
 
 /**
  * @author Filip Hrisafov
  */
-public class Target {
+public class Source {
 
     private String value;
