+++ processor/src/test/java/org/mapstruct/ap/test/bugs/_2807/spring/Issue2807Test.java
@@ -3,11 +3,11 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.bugs._2807;
+package org.mapstruct.ap.test.bugs._2807.spring;
 
-import org.mapstruct.ap.test.bugs._2807.after.AfterMethod;
-import org.mapstruct.ap.test.bugs._2807.before.BeforeMethod;
-import org.mapstruct.ap.test.bugs._2807.beforewithtarget.BeforeWithTarget;
+import org.mapstruct.ap.test.bugs._2807.spring.after.AfterMethod;
+import org.mapstruct.ap.test.bugs._2807.spring.before.BeforeMethod;
+import org.mapstruct.ap.test.bugs._2807.spring.beforewithtarget.BeforeWithTarget;
 import org.mapstruct.ap.testutil.IssueKey;
 import org.mapstruct.ap.testutil.ProcessorTest;
 import org.mapstruct.ap.testutil.WithClasses;
