+++ processor/src/test/java/org/mapstruct/ap/test/bugs/_2807/spring/SpringLifeCycleMapper.java
@@ -3,15 +3,15 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.bugs._2807;
+package org.mapstruct.ap.test.bugs._2807.spring;
 
 import java.util.List;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
-import org.mapstruct.ap.test.bugs._2807.after.AfterMethod;
-import org.mapstruct.ap.test.bugs._2807.before.BeforeMethod;
-import org.mapstruct.ap.test.bugs._2807.beforewithtarget.BeforeWithTarget;
+import org.mapstruct.ap.test.bugs._2807.spring.after.AfterMethod;
+import org.mapstruct.ap.test.bugs._2807.spring.before.BeforeMethod;
+import org.mapstruct.ap.test.bugs._2807.spring.beforewithtarget.BeforeWithTarget;
 import org.mapstruct.factory.Mappers;
 
 /**
