+++ processor/src/test/java/org/mapstruct/ap/test/bugs/_2807/spring/before/BeforeMethod.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.bugs._2807.before;
+package org.mapstruct.ap.test.bugs._2807.spring.before;
 
 import org.mapstruct.BeforeMapping;
