+++ processor/src/test/java/org/mapstruct/ap/test/bugs/_2807/spring/beforewithtarget/BeforeWithTarget.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.bugs._2807.beforewithtarget;
+package org.mapstruct.ap.test.bugs._2807.spring.beforewithtarget;
 
 import java.util.List;
