+++ processor/src/test/java/org/mapstruct/ap/test/bugs/_880/spring/Poodle.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.bugs._880;
+package org.mapstruct.ap.test.bugs._880.spring;
 
 /**
  * @author Andreas Gudian
